Renames serializer.mjs to yaml.mjs
