+++ workspaces/resource-optimization/plugins/resource-optimization-common/scripts/lib/yaml.mjs
